requirements.txt
flask
transformers
torch
wikipedia
google-genai
requests